Set the original strings if missing in PO
Set the original strings if missing translations in PO and converting it with --keyasareference option.
Unlike an usual gettext applications, this sort of apps expects to see strings in JSON regardless of i18n.
So if JSON is missing strings, that may ends up to show no strings.
This change will fixes such issue.