Fix: `--compatibilityJSON` correctly passed as str
Previously, the CLI option would not accept a value and hence would be
interpreted as a boolean switch, which resulted in the program not
passing the correct value to the converter and resulting in incorrect PO
file output.

This seems to be due to the way value vs. bool options are delcared,
see: https://github.com/tj/commander.js#common-option-types-boolean-and-value