Do not require -l if -t is present
When converting from .po to .json, you currently get an error if
specifiying -s and -t but not -l arguments on the command line.
However, -l is only used to build the path for the target and not
present in the json output.

This updates the logic to require -s and -t OR -l.